# ISSUE-4345 integrate fuse table with new processor