--- a/query/src/pipelines/new/executor/executor_worker_context.rs
+++ b/query/src/pipelines/new/executor/executor_worker_context.rs
@@ -88,11 +88,13 @@ impl ExecutorWorkerContext {
         executor: &PipelineExecutor,
     ) -> Result<Option<NodeIndex>> {
         let worker_id = self.worker_num;
+        let workers_notify = self.get_workers_notify().clone();
         let tasks_queue = executor.global_tasks_queue.clone();
         executor.async_runtime.spawn(async move {
             let res = processor.async_process().await;
             let task = CompletedAsyncTask::create(processor, worker_id, res);
             tasks_queue.completed_async_task(task);
+            workers_notify.wakeup(worker_id);
         });
 
         Ok(None)