--- a/query/src/sessions/query_ctx.rs
+++ b/query/src/sessions/query_ctx.rs
@@ -344,7 +344,7 @@ impl QueryContext {
     }
 
     // Get the storage data accessor operator from the session manager.
-    pub async fn get_storage_operator(&self) -> Result<Operator> {
+    pub fn get_storage_operator(&self) -> Result<Operator> {
         let operator = self.shared.session.get_storage_operator();
         Ok(operator.layer(self.shared.dal_ctx.as_ref().clone()))
     }