--- a/query/src/storages/fuse/io/block_reader.rs
+++ b/query/src/storages/fuse/io/block_reader.rs
@@ -38,6 +38,7 @@ use futures::AsyncReadExt;
 use futures::StreamExt;
 use futures::TryStreamExt;
 use opendal::Operator;
+use opendal::Reader;
 
 use crate::sessions::QueryContext;
 use crate::storages::fuse::fuse_part::ColumnMeta;
@@ -152,6 +153,77 @@ impl BlockReader {
         }
 
         Ok((rows, columns_array_iter))
+    }
+
+    pub fn deserialize(&self, part: PartInfoPtr, chunks: Vec<Vec<u8>>) -> Result<DataBlock> {
+        if self.projection.len() != chunks.len() {
+            return Err(ErrorCode::LogicalError(
+                "Columns chunk len must be equals projections len.",
+            ));
+        }
+
+        let part = FusePartInfo::from_part(&part)?;
+        let mut columns_array_iter = Vec::with_capacity(self.projection.len());
+
+        let num_rows = part.nums_rows;
+        for (index, column_chunk) in chunks.into_iter().enumerate() {
+            let index = self.projection[index];
+            let field = self.arrow_schema.fields[index].clone();
+            let column_descriptor = self.parquet_schema_descriptor.column(index);
+            let column_meta = &part.columns_meta[&index];
+            columns_array_iter.push(Self::to_deserialize(
+                column_meta,
+                column_chunk,
+                num_rows,
+                column_descriptor,
+                field,
+            )?);
+        }
+
+        let mut deserializer = RowGroupDeserializer::new(columns_array_iter, num_rows, None);
+
+        match deserializer.next() {
+            None => Err(ErrorCode::ParquetError("fail to get a chunk")),
+            Some(Err(cause)) => Err(ErrorCode::from(cause)),
+            Some(Ok(chunk)) => DataBlock::from_chunk(&self.projected_schema, &chunk),
+        }
+    }
+
+    pub async fn read_columns_data(&self, part: PartInfoPtr) -> Result<Vec<Vec<u8>>> {
+        let part = FusePartInfo::from_part(&part)?;
+        let mut join_handlers = Vec::with_capacity(self.projection.len());
+
+        for index in &self.projection {
+            let column_meta = &part.columns_meta[index];
+
+            let column_reader = self
+                .operator
+                .object(&part.location)
+                .range_reader(column_meta.offset, column_meta.length);
+
+            let column_chunk = vec![0; column_meta.length as usize];
+            join_handlers.push(Self::read_column(column_reader, column_chunk));
+        }
+
+        futures::future::try_join_all(join_handlers).await
+    }
+
+    async fn read_column(mut column_reader: Reader, mut chunk: Vec<u8>) -> Result<Vec<u8>> {
+        let handler = common_base::tokio::spawn(async move {
+            tracing::debug!("read_exact | Begin, {:?}", std::thread::current());
+            column_reader.read_exact(&mut chunk).await?;
+            tracing::debug!("read_exact | End, {:?}", std::thread::current());
+            Result::Ok(chunk)
+        });
+
+        match handler.await {
+            Ok(Ok(data)) => Ok(data),
+            Ok(Err(cause)) => Err(cause),
+            Err(cause) => Err(ErrorCode::TokioError(format!(
+                "Cannot join future {:?}",
+                cause
+            ))),
+        }
     }
 
     #[tracing::instrument(level = "debug", skip_all)]