--- a/query/src/storages/fuse/operations/commit.rs
+++ b/query/src/storages/fuse/operations/commit.rs
@@ -163,7 +163,7 @@ impl FuseTable {
         let uuid = new_snapshot.snapshot_id;
         let snapshot_loc = self.meta_locations().snapshot_location_from_uuid(&uuid);
         let bytes = serde_json::to_vec(&new_snapshot)?;
-        let operator = ctx.get_storage_operator().await?;
+        let operator = ctx.get_storage_operator()?;
         operator
             .object(&snapshot_loc)
             .writer()