--- a/query/src/storages/fuse/operations/read.rs
+++ b/query/src/storages/fuse/operations/read.rs
@@ -13,10 +13,10 @@
 //  limitations under the License.
 //
 
-use std::future::Future;
 use std::sync::Arc;
 
 use common_datablocks::DataBlock;
+use common_exception::ErrorCode;
 use common_exception::Result;
 use common_planners::Extras;
 use common_planners::PartInfoPtr;
@@ -26,13 +26,14 @@ use common_tracing::tracing_futures::Instrument;
 use futures::StreamExt;
 
 use crate::pipelines::new::processors::port::OutputPort;
+use crate::pipelines::new::processors::processor::Event;
 use crate::pipelines::new::processors::processor::ProcessorPtr;
-use crate::pipelines::new::processors::AsyncSource;
-use crate::pipelines::new::processors::AsyncSourcer;
+use crate::pipelines::new::processors::Processor;
 use crate::pipelines::new::NewPipeline;
 use crate::pipelines::new::SourcePipeBuilder;
 use crate::sessions::QueryContext;
 use crate::storages::fuse::io::BlockReader;
+use crate::storages::fuse::operations::read::State::Generated;
 use crate::storages::fuse::FuseTable;
 
 impl FuseTable {
@@ -112,10 +113,18 @@ impl FuseTable {
     }
 }
 
+enum State {
+    ReadData,
+    Deserialize(Vec<Vec<u8>>),
+    Generated(DataBlock),
+    Finish,
+}
+
 struct FuseTableSource {
-    finish: bool,
+    state: State,
     part: PartInfoPtr,
     block_reader: Arc<BlockReader>,
+    output: Arc<OutputPort>,
 }
 
 impl FuseTableSource {
@@ -124,29 +133,67 @@ impl FuseTableSource {
         block_reader: Arc<BlockReader>,
         part: PartInfoPtr,
     ) -> Result<ProcessorPtr> {
-        AsyncSourcer::create(output, FuseTableSource {
-            finish: false,
+        Ok(ProcessorPtr::create(Box::new(FuseTableSource {
+            state: State::ReadData,
             part,
             block_reader,
-        })
+            output,
+        })))
     }
 }
 
-impl AsyncSource for FuseTableSource {
-    const NAME: &'static str = "FuseEngineSource";
+#[async_trait::async_trait]
+impl Processor for FuseTableSource {
+    fn name(&self) -> &'static str {
+        "FuseEngineSource"
+    }
 
-    type BlockFuture<'a>
-    where Self: 'a
-    = impl Future<Output = Result<Option<DataBlock>>>;
-
-    fn generate(&mut self) -> Self::BlockFuture<'_> {
-        async {
-            if self.finish {
-                return Ok(None);
-            }
-
-            self.finish = true;
-            Ok(Some(self.block_reader.read(self.part.clone()).await?))
+    fn event(&mut self) -> Result<Event> {
+        if matches!(self.state, State::Finish) {
+            self.output.finish();
+            return Ok(Event::Finished);
         }
+
+        if self.output.is_finished() {
+            return Ok(Event::Finished);
+        }
+
+        if !self.output.can_push() {
+            return Ok(Event::NeedConsume);
+        }
+
+        if matches!(self.state, State::Generated(_)) {
+            println!("Generated");
+            if let Generated(data_block) = std::mem::replace(&mut self.state, State::Finish) {
+                println!("Generated {:?}", data_block);
+                self.output.push_data(Ok(data_block));
+                return Ok(Event::NeedConsume);
+            }
+        }
+
+        match self.state {
+            State::Finish => Ok(Event::Finished),
+            State::ReadData => Ok(Event::Async),
+            State::Deserialize(_) => Ok(Event::Sync),
+            State::Generated(_) => Err(ErrorCode::LogicalError("It's a bug.")),
+        }
+    }
+
+    fn process(&mut self) -> Result<()> {
+        if let State::Deserialize(chunks) = std::mem::replace(&mut self.state, State::ReadData) {
+            let part = self.part.clone();
+            self.state = State::Generated(self.block_reader.deserialize(part, chunks)?);
+        }
+
+        Ok(())
+    }
+
+    async fn async_process(&mut self) -> Result<()> {
+        if let State::ReadData = self.state {
+            let part = self.part.clone();
+            self.state = State::Deserialize(self.block_reader.read_columns_data(part).await?);
+        }
+
+        Ok(())
     }
 }