--- a/query/tests/it/sessions/query_ctx.rs
+++ b/query/tests/it/sessions/query_ctx.rs
@@ -35,7 +35,7 @@ async fn test_get_storage_accessor_s3() -> Result<()> {
 
     let qctx = crate::tests::create_query_context_with_config(conf).await?;
 
-    let _ = qctx.get_storage_operator().await?;
+    let _ = qctx.get_storage_operator()?;
 
     Ok(())
 }
@@ -52,7 +52,7 @@ async fn test_get_storage_accessor_fs() -> Result<()> {
 
     let qctx = crate::tests::create_query_context_with_config(conf).await?;
 
-    let _ = qctx.get_storage_operator().await?;
+    let _ = qctx.get_storage_operator()?;
 
     Ok(())
 }