I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

integrate fuse table with new processor

## Changelog

- New Feature

## Related Issues

Fixes #4345

## Test Plan

Unit Tests

Stateless Tests