@BohuTANG 2022-03-16
:tada: 

I have test [ontime on s3](https://databend.rs/doc/performance/ec2-s3-performance), there is a bit performance improve.